requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==21.2.0
tensorflow-cpu==2.12.0
h5py==3.8.0
numpy==1.23.5
requests>=2.28
